% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_sample.R
\name{spatial_sample}
\alias{spatial_sample}
\title{Spatial sample}
\usage{
spatial_sample(
  data,
  n,
  design,
  r,
  bias_var,
  bias_thresh,
  strata_var,
  clh_var,
  clh_iter
)
}
\arguments{
\item{data}{A \code{data.frame} object, ideally the output from
spatialsample_prep().}

\item{n}{The number of \code{data.frame} rows to select (i.e., the sample size).}

\item{design}{Which sampling design should be used to select the rows?
\itemize{
\item \code{"random"} (the default): random sampling via \code{dplyr::slice_sample()}.
\item \code{"biased"}: random sampling of rows for which \code{bias_var} value is greater than
\code{bias_thresh} via \code{dplyr::slice_sample()}.
\item \code{"stratified"}: stratified random sampling. Randomly select (n / number of strata) rows for each
value of \code{strata_var} via \code{dplyr::slice_sample()}.
\item \code{"clh"}: conditioned latin hypercube sampling via \code{clhs::clhs()}.
\item \code{"balanced"}: spatially balanced sampling via \code{MBHdesign::quasiSamp()}.
\item \code{"balanced_stratified"}: spatially balanced stratified sampling via
\code{MBHdesign::quasiSamp()}.
}}

\item{r}{A \code{SpatRaster} object specifying the}

\item{bias_var}{description}

\item{bias_thresh}{description}

\item{strata_var}{description}

\item{clh_var}{description}

\item{clh_iter}{A positive number, giving the number of iterations for the
Metropolis-Hastings annealing process. If \code{design} is \code{"clh"}, this value is
passed to the \code{iter} argument of \code{clhs::clhs()}.}
}
\value{
description
}
\description{
Take a sample of the data according to the specified sampling design
}
