% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_sample.R
\name{spatialsample_prep}
\alias{spatialsample_prep}
\title{Spatial sample prep}
\usage{
spatialsample_prep(x, cells = FALSE, xy = FALSE, scale = FALSE)
}
\arguments{
\item{data}{description}

\item{n}{description}

\item{design}{description}
}
\value{
description
}
\description{
Spatial sample prep
}
