% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_sample.R
\name{spatialsample_random}
\alias{spatialsample_random}
\title{Spatial sample of data}
\usage{
spatialsample_random(data, n)
}
\arguments{
\item{data}{A \code{data.frame} object, ideally the output from
spatialsample_prep().}

\item{n}{The number of \code{data.frame} rows to select (i.e., the sample size).}

\item{var}{description}
}
\value{
description
}
\description{
Spatial sample of data
}
