% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify}
\usage{
stratify(x, n_strata, equal_split = TRUE, probs = NULL, vals = NULL)
}
\arguments{
\item{x}{Either a numeric vector or a \code{SpatRaster} object. If \code{SpatRaster},
only the first layer will be used.}

\item{n_strata}{Number of strata.}

\item{equal_split}{Whether or not to stratify into equally-spaced quantiles
of the data values. If \code{TRUE} (the default), \code{probs} and \code{vals} should both
be \code{NULL}.}

\item{probs}{A numeric vector of probabilities, sorted increasingly, for the
cut points with values between 0 and 1. The length of this vector should equal
\code{n_strata} - 1.}

\item{vals}{A numeric vector of break point values, sorted increasingly. The
length of this vector should equal \code{n_strata} - 1.}
}
\value{
An object of the same type as \code{x} (either a numeric vector or a
\code{SpatRaster} object).
}
\description{
Stratify a vector or \code{SpatRaster} object into the given number of strata.
}
